Change Aurora users to superusers.
This view now shows people with the specified roles *in Northstar*,
rather than their local accounts.
